Make it work for really big arrays too

+++ intersection.pyx
@@ -45,14 +45,14 @@ def intersection(arr1, start1, length1, arr2, start2, length2):
 
     cdef const unsigned char* in1 = &buf1[0]
     cdef const unsigned char* in2 = &buf2[0]
-    cdef int len1 = buf1.nbytes
-    cdef int len2 = buf2.nbytes
+    cdef size_t len1 = buf1.nbytes
+    cdef size_t len2 = buf2.nbytes
 
     out = <char*>malloc(max(len1, len2))
 
-    cdef int i = 0
-    cdef int j = 0
-    cdef int k = 0
+    cdef size_t i = 0
+    cdef size_t j = 0
+    cdef size_t k = 0
 
     while i < len1 and j < len2:
         x = read_bytes(in1+i, size)
@@ -73,14 +73,14 @@ def intersection(arr1, start1, length1, arr2, start2, length2):
 cdef extern from "string.h":
     void *memcpy(void *dest, const void *src, size_t len)
 
-cdef inline long read_bytes(const void *ptr, int size):
+cdef inline size_t read_bytes(const void *ptr, int size):
     """Read an integer of the given number of bytes from a pointer."""
 
-    cdef long result = 0
+    cdef size_t result = 0
     memcpy(&result, ptr, size)
     return result
 
-cdef inline void write_bytes(void *ptr, long value, int size):
+cdef inline void write_bytes(void *ptr, size_t value, int size):
     """Write an integer of the given number of bytes to a pointer."""
 
     memcpy(ptr, &value, size)
